inject tests has been rewrote
